+++ NEWS.md
@@ -2,7 +2,7 @@
 
 - Added function 'getUnisensSignalSampleCount'
 - Added functionality to specify start and end index of signal entry reading
-- Added funtionality to read signal entries in chunks 
+- Added functionality to read signal entries in chunks 
 
 # v0.3.2 (2020-01-16)